Yes and no Quiz game
So, I made this basic yes-and-no quiz game to get the hang of user input and method creation. Along the way, I figured out how to use the ToLower method and even threw in a scoring system. Sure, there might be better ways to do it, but hey, it gets the job done